requirements.txt
nbformat
pandas
openai>=1.50.0
python-dotenv
transformers
torch --extra-index-url https://download.pytorch.org/whl/cu118
torchvision --extra-index-url https://download.pytorch.org/whl/cu118
torchaudio --extra-index-url https://download.pytorch.org/whl/cu118
tenacity
chromadb
langchain
langchain-openai
unstructured
langchain-community
unstructured[md]
streamlit
seaborn
